fix: rename jwt auth file
